feat: moved pdf client-functionality out of page.tsx and added page title templates!
